Backend/requirements.txt
baostock==0.8.9
pandas>=2.2.0
psycopg[binary]>=3.2.0
python-dotenv==1.0.0
akshare>=1.12.50
fastapi==0.115.6
uvicorn[standard]==0.34.0
pypinyin==0.54.0
chinesecalendar>=1.9.0